requirements.txt
openai
requests
python-dotenv
firecrawl-py
syncsdk
elevenlabs
ffmpeg-python

